mobile/requirements.txt
pytest>=7.3.1
pyyaml>=6.0.0
pydantic>=2.0
uiautomator2>=3.0.0
facebook-wda>=1.3.0
hmdriver2>=0.3.0
pillow>=10.0.0
opencv-python-headless>=4.8.0
adbutils>=2.0.0
loguru>=0.7.0
